disdrodl/requirements.txt
pyserial==3.5
dataclasses==0.8
loguru==0.5.3